test existence of file before deleting it

+++ aes/create-in-vpc-with-client/cleanup.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
-rm ./connect.sh
+[ -f ./connect.sh ] && rm ./connect.sh
 aws cloudformation delete-stack --stack-name Valentine-Test-AES-Stack
